+++ CS_490_Project_2/src/Business_Logic/Review.java
@@ -11,17 +11,17 @@ package Business_Logic;
  */
 public class Review {
     //Class variables
-    private int rating;
+    private double rating;
     private String review;
     
     //Class constructor
-    public Review (int Rating, String Review) {
+    public Review (double Rating, String Review) {
         this.rating = Rating;
         this.review = Review;
     }
 
     //Class get fucntions
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
